[collect] Do not compress top-level collector archive
Previously, when running `sos collect`, the resulting archive that
contained sos report archives, would itself be compressed. This "top
level" compression did not actually gain us much, but did have a time
penalty to overall collect execution times.

This commit solves this by allowing a `None` value to be
used for the compression method when calling `Archive.finalize()`, which
will skip compressing that archive. The actual sos reports that a
collect execution captures are still compressed.

Note: this "none" option is not currently exposed on the CLI, and all
sos reports will still be compressed.

Resolves: #3992

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>