[cirrus] Temporarily disable Fedora testing
We have been seeing some odd failures with the CI around Fedora,
particularly with the rpmbuild failing which then cancels the rest of
the suite due to dependency resolution.

We're working with Cirrus support, but until we have a fix let's disable
the Fedora testing to unblock the rest of the suite.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>