[tests] Fix expected value for preset tests
The default value for networking.namespaces in the file
tests/report_tests/options_tests/options_tests_sos.conf
is 100, while the value for this variable in
tests/report_tests/options_tests/options_tests_preset.json
is 200. To make sure that preset beats the config file
we need to change the value from 100 to 200.

Related: #4087

Signed-off-by: Jose Castillo <jcastillo@redhat.com>